+++ internal/observability/graph.go
@@ -49,7 +49,10 @@ type directionName string
 const (
 	_directionOutbound directionName = "outbound"
 	_directionInbound  directionName = "inbound"
-	_redactedTagVal                  = "__dropped__"
+
+	// _droppedTagValue represents the value of a metric tag when the tag
+	// is being blocked.
+	_droppedTagValue = "__dropped__"
 )
 
 // A graph represents a collection of services: each service is a node, and we
@@ -203,7 +206,7 @@ func newEdge(logger *zap.Logger, meter *metrics.Scope, metricTagsBlocklist []str
 	}
 
 	for _, tagName := range metricTagsBlocklist {
-		tags[tagName] = _redactedTagVal
+		tags[tagName] = _droppedTagValue
 	}
 
 	// metrics for all RPCs